Fix generated project with wrong duplicated package name Closes #13581
