+++ devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/commandmode-example/kotlin/src/main/kotlin/org/acme/{main.class-name}.tpl.qute.kt
@@ -1,0 +1,17 @@
+package org.acme
+
+import javax.inject.Inject
+
+import io.quarkus.runtime.QuarkusApplication
+import io.quarkus.runtime.annotations.QuarkusMain
+
+@QuarkusMain
+class {main.class-name}: QuarkusApplication {
+
+    override fun run(vararg args: String?): Int {
+        val name =  if (args.isNotEmpty()) args.joinToString(",") else "{greeting.default-name}"
+        println(name)
+        return 0
+    }
+
+}
