--- devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/commandmode-example/kotlin/src/main/kotlin/{package-name.dir}/{main.class-name}.tpl.qute.kt
@@ -1,17 +1,0 @@
-package {package-name}
-
-import javax.inject.Inject
-
-import io.quarkus.runtime.QuarkusApplication
-import io.quarkus.runtime.annotations.QuarkusMain
-
-@QuarkusMain
-class {main.class-name}: QuarkusApplication {
-
-    override fun run(vararg args: String?): Int {
-        val name =  if (args.isNotEmpty()) args.joinToString(",") else "{greeting.default-name}"
-        println(name)
-        return 0
-    }
-
-}
