+++ devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/resteasy-example/codestart.yml
@@ -12,7 +12,6 @@ language:
         class-name: GreetingResource
         path: "/hello-resteasy"
         response: "Hello RESTEasy"
-      package-name: org.acme
       guide: https://quarkus.io/guides/rest-json
     dependencies:
       - io.quarkus:quarkus-resteasy
