--- devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/resteasy-example/java/src/main/java/{package-name.dir}/{resource.class-name}.tpl.qute.java
@@ -1,16 +1,0 @@
-package {package-name};
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-
-@Path("{resource.path}")
-public class {resource.class-name} {
-
-    @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    public String hello() {
-        return "{resource.response}";
-    }
-}
