+++ devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/resteasy-example/java/src/native-test/java/org/acme/Native{resource.class-name}IT.tpl.qute.java
@@ -1,0 +1,9 @@
+package org.acme;
+
+import io.quarkus.test.junit.NativeImageTest;
+
+@NativeImageTest
+public class Native{resource.class-name}IT extends {resource.class-name}Test {
+
+    // Execute the same tests but in native mode.
+}
