--- devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/resteasy-example/kotlin/src/main/kotlin/{package-name.dir}/{resource.class-name}.tpl.qute.kt
@@ -1,14 +1,0 @@
-package {package-name}
-
-import javax.ws.rs.GET
-import javax.ws.rs.Path
-import javax.ws.rs.Produces
-import javax.ws.rs.core.MediaType
-
-@Path("{resource.path}")
-class {resource.class-name} {
-
-    @GET
-    @Produces(MediaType.TEXT_PLAIN)
-    fun hello() = "{resource.response}"
-}
