--- devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/resteasy-example/kotlin/src/native-test/kotlin/{package-name.dir}/Native{resource.class-name}IT.tpl.qute.kt
@@ -1,6 +1,0 @@
-package {package-name}
-
-import io.quarkus.test.junit.NativeImageTest
-
-@NativeImageTest
-class Native{resource.class-name}IT : {resource.class-name}Test()
