--- devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/resteasy-example/scala/src/main/scala/{package-name.dir}/{resource.class-name}.tpl.qute.scala
@@ -1,12 +1,0 @@
-package {package-name}
-
-import javax.ws.rs.\{GET, Path, Produces}
-import javax.ws.rs.core.MediaType
-
-@Path("{resource.path}")
-class {resource.class-name} {
-
-    @GET
-    @Produces(Array[String](MediaType.TEXT_PLAIN))
-    def hello() = "{resource.response}"
-}
