+++ devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/spring-web-example/java/src/main/java/org/acme/{resource.class-name}.tpl.qute.java
@@ -1,0 +1,15 @@
+package org.acme;
+
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("{resource.path}")
+public class {resource.class-name} {
+
+    @GetMapping
+    public String hello() {
+        return "{resource.response}";
+    }
+}
