--- devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/spring-web-example/java/src/main/java/{package-name.dir}/{resource.class-name}.tpl.qute.java
@@ -1,15 +1,0 @@
-package {package-name};
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("{resource.path}")
-public class {resource.class-name} {
-
-    @GetMapping
-    public String hello() {
-        return "{resource.response}";
-    }
-}
