--- devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/spring-web-example/java/src/native-test/java/{package-name.dir}/Native{resource.class-name}IT.tpl.qute.java
@@ -1,9 +1,0 @@
-package {package-name};
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class Native{resource.class-name}IT extends {resource.class-name}Test {
-
-    // Execute the same tests but in native mode.
-}
