--- devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/spring-web-example/kotlin/src/main/kotlin/{package-name.dir}/{resource.class-name}.tpl.qute.kt
@@ -1,14 +1,0 @@
-package {package-name}
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-
-@RestController
-@RequestMapping("{resource.path}")
-class {resource.class-name} {
-
-    @GetMapping
-    fun hello() = "{resource.response}"
-}
