+++ devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/spring-web-example/kotlin/src/native-test/kotlin/org/acme/Native{resource.class-name}IT.tpl.qute.kt
@@ -1,0 +1,6 @@
+package org.acme
+
+import io.quarkus.test.junit.NativeImageTest
+
+@NativeImageTest
+class Native{resource.class-name}IT : {resource.class-name}Test()
