+++ devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/spring-web-example/scala/src/main/scala/org/acme/{resource.class-name}.tpl.qute.scala
@@ -1,0 +1,12 @@
+package org.acme;
+
+import org.springframework.web.bind.annotation.\{GetMapping, RequestMapping, RestController}
+
+
+@RestController
+@RequestMapping(Array[String]("{resource.path}"))
+class {resource.class-name} {
+
+    @GetMapping
+    def hello() = "{resource.response}"
+}
