+++ devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/spring-web-example/scala/src/native-test/scala/org/acme/Native{resource.class-name}IT.tpl.qute.scala
@@ -1,0 +1,6 @@
+package org.acme
+
+import io.quarkus.test.junit.NativeImageTest
+
+@NativeImageTest
+class Native{resource.class-name}IT extends {resource.class-name}Test
