--- devtools/platform-descriptor-json/src/main/resources/codestarts/quarkus/core/examples/spring-web-example/scala/src/native-test/scala/{package-name.dir}/Native{resource.class-name}IT.tpl.qute.scala
@@ -1,6 +1,0 @@
-package {package-name}
-
-import io.quarkus.test.junit.NativeImageTest
-
-@NativeImageTest
-class Native{resource.class-name}IT extends {resource.class-name}Test
