+++ independent-projects/tools/devtools-common/src/main/java/io/quarkus/devtools/commands/handlers/CreateProjectCommandHandler.java
@@ -8,6 +8,7 @@ import static io.quarkus.devtools.commands.handlers.QuarkusCommandHandlers.compu
 import static io.quarkus.devtools.project.codegen.ProjectGenerator.BOM_ARTIFACT_ID;
 import static io.quarkus.devtools.project.codegen.ProjectGenerator.BOM_GROUP_ID;
 import static io.quarkus.devtools.project.codegen.ProjectGenerator.BOM_VERSION;
+import static io.quarkus.devtools.project.codegen.ProjectGenerator.CLASS_NAME;
 import static io.quarkus.devtools.project.codegen.ProjectGenerator.PACKAGE_NAME;
 import static io.quarkus.devtools.project.codegen.ProjectGenerator.PROJECT_GROUP_ID;
 import static io.quarkus.devtools.project.codegen.ProjectGenerator.QUARKUS_VERSION;
@@ -59,10 +60,17 @@ public class CreateProjectCommandHandler implements QuarkusCommandHandler {
         });
 
         // Default to cleaned groupId if packageName not set
+        final String className = invocation.getStringValue(CLASS_NAME);
         final String pkgName = invocation.getStringValue(PACKAGE_NAME);
         final String groupId = invocation.getStringValue(PROJECT_GROUP_ID);
-        if (pkgName == null && groupId != null) {
-            invocation.setValue(PACKAGE_NAME, groupId.replace("-", ".").replace("_", "."));
+        if (pkgName == null) {
+            if (className != null && className.contains(".")) {
+                final int idx = className.lastIndexOf('.');
+                invocation.setValue(PACKAGE_NAME, className.substring(0, idx));
+                invocation.setValue(CLASS_NAME, className.substring(idx + 1));
+            } else if (groupId != null) {
+                invocation.setValue(PACKAGE_NAME, groupId.replace("-", ".").replace("_", "."));
+            }
         }
 
         final List<AppArtifactCoords> extensionsToAdd = computeCoordsFromQuery(invocation, extensionsQuery);
