+++ integration-tests/gradle/src/test/java/io/quarkus/gradle/QuarkusPluginFunctionalTest.java
@@ -88,7 +88,7 @@ public class QuarkusPluginFunctionalTest extends QuarkusGradleWrapperTestBase {
         BuildResult firstBuild = runGradleWrapper(projectRoot, "quarkusBuild", "--stacktrace");
         assertThat(firstBuild.getTasks().get(":quarkusBuild")).isEqualTo(BuildResult.SUCCESS_OUTCOME);
 
-        final File greetingResourceFile = projectRoot.toPath().resolve("src/main/java/com/acme/foo/GreetingResource.java")
+        final File greetingResourceFile = projectRoot.toPath().resolve("src/main/java/org/acme/foo/GreetingResource.java")
                 .toFile();
         DevModeTestUtils.filter(greetingResourceFile, ImmutableMap.of("\"/greeting\"", "\"/test/hello\""));
 
@@ -170,7 +170,7 @@ public class QuarkusPluginFunctionalTest extends QuarkusGradleWrapperTestBase {
                         .artifactId("foo")
                         .version("1.0.0-SNAPSHOT")
                         .buildTool(BuildTool.GRADLE)
-                        .className("org.acme.GreetingResource")
+                        .className("org.acme.foo.GreetingResource")
                         .sourceType(sourceType)
                         .doCreateProject(context))
                                 .withFailMessage("Project was not created")
